# membuffer: compare the keys of ART by chunk