--- a/internal/unionstore/art/art_iterator.go
+++ b/internal/unionstore/art/art_iterator.go
@@ -360,8 +360,8 @@ func (it *baseIter) next() artNode {
 			if idx >= 0 && idx < int(n4.nodeNum) {
 				it.idxes[depth] = idx
 				child = &n4.children[idx]
-			} else if idx == int(n4.nodeNum) {
-				// idx == n4.nodeNum means this node is drain, break to pop stack.
+			} else if idx >= int(n4.nodeNum) {
+				// idx >= n4.nodeNum means this node is drain, break to pop stack.
 				break
 			} else {
 				panicForInvalidIndex(idx)
@@ -380,8 +380,8 @@ func (it *baseIter) next() artNode {
 			if idx >= 0 && idx < int(n16.nodeNum) {
 				it.idxes[depth] = idx
 				child = &n16.children[idx]
-			} else if idx == int(n16.nodeNum) {
-				// idx == n16.nodeNum means this node is drain, break to pop stack.
+			} else if idx >= int(n16.nodeNum) {
+				// idx >= n16.nodeNum means this node is drain, break to pop stack.
 				break
 			} else {
 				panicForInvalidIndex(idx)