--- a/internal/unionstore/art/art_node.go
+++ b/internal/unionstore/art/art_node.go
@@ -16,7 +16,6 @@ package art
 
 import (
 	"bytes"
-	"encoding/binary"
 	"math"
 	"math/bits"
 	"runtime"
@@ -59,22 +58,6 @@ const (
 	node256size = int(unsafe.Sizeof(node256{}))
 	leafSize    = int(unsafe.Sizeof(artLeaf{}))
 )
-
-var nativeEndian binary.ByteOrder
-
-func init() {
-	buf := [2]byte{}
-	*(*uint16)(unsafe.Pointer(&buf[0])) = uint16(0xABCD)
-
-	switch buf {
-	case [2]byte{0xCD, 0xAB}:
-		nativeEndian = binary.LittleEndian
-	case [2]byte{0xAB, 0xCD}:
-		nativeEndian = binary.BigEndian
-	default:
-		panic("Could not determine native endianness.")
-	}
-}
 
 var nullArtNode = artNode{kind: typeInvalid, addr: arena.NullAddr}