--- a/internal/unionstore/art/art_node.go
+++ b/internal/unionstore/art/art_node.go
@@ -346,14 +346,14 @@ func (an *artNode) asNode256(a *artAllocator) *node256 {
 	return a.getNode256(an.addr)
 }
 
+// for amd64 and arm64 architectures, we use the chunk comparison to speed up finding the longest common prefix.
+const enableChunkComparison = runtime.GOARCH == "amd64" || runtime.GOARCH == "arm64"
+
 // longestCommonPrefix returns the length of the longest common prefix of two keys.
 // the LCP is calculated from the given depth, you need to guarantee l1Key[:depth] equals to l2Key[:depth] before calling this function.
 func longestCommonPrefix(l1Key, l2Key artKey, depth uint32) uint32 {
-	switch runtime.GOARCH {
-	case "amd64", "arm64":
-		// for amd64 and arm64 architectures, we use the chunk comparison to accelerate the calculation.
+	if enableChunkComparison {
 		return longestCommonPrefixByChunk(l1Key, l2Key, depth)
-	default:
 	}
 	// For other architectures, we use the byte-by-byte comparison.
 	idx, limit := depth, uint32(min(len(l1Key), len(l2Key)))