ref pingcap/tidb#55287

In order to perform path compression in ART, we need to find the longest common prefix the keys in many places, when the common prefix is long, this can be slow.

This PR compares the keys by chunk. Because a chunk is a uint64 which can contains 8 bytes, this can improve the performance a lot when the common prefix is long.

- Long Common Prefix
```
# before
BenchmarkMemBufferSetGetLongKey
BenchmarkMemBufferSetGetLongKey/ART
BenchmarkMemBufferSetGetLongKey/ART-32           1000000               874.3 ns/op          2149 B/op          0 allocs/op

# this PR
BenchmarkMemBufferSetGetLongKey
BenchmarkMemBufferSetGetLongKey/ART
BenchmarkMemBufferSetGetLongKey/ART-32           1000000               524.5 ns/op          2149 B/op          0 allocs/op
```

- Common Cases

```
# before
BenchmarkPut
BenchmarkPut/ART
BenchmarkPut/ART-32              1000000                91.70 ns/op          335 B/op          0 allocs/op
BenchmarkPutRandom
BenchmarkPutRandom/ART
BenchmarkPutRandom/ART-32                1000000               182.1 ns/op           404 B/op          0 allocs/op
BenchmarkGet
BenchmarkGet/ART
BenchmarkGet/ART-32                      1000000                31.31 ns/op            0 B/op          0 allocs/op
BenchmarkGetRandom
BenchmarkGetRandom/ART
BenchmarkGetRandom/ART-32                1000000               142.7 ns/op    

# this PR
BenchmarkPut
BenchmarkPut/ART
BenchmarkPut/ART-32              1000000                89.18 ns/op          335 B/op          0 allocs/op
BenchmarkPutRandom
BenchmarkPutRandom/ART
BenchmarkPutRandom/ART-32                1000000               175.8 ns/op           404 B/op          0 allocs/op
BenchmarkGet
BenchmarkGet/ART
BenchmarkGet/ART-32                      1000000                31.74 ns/op            0 B/op          0 allocs/op
BenchmarkGetRandom
BenchmarkGetRandom/ART
BenchmarkGetRandom/ART-32                1000000               132.5 ns/op   
```