@you06 2024-10-30
> Would you like to compare it with something like this?

The unsafe implementation.

- Long Common Prefix

That's about +15% speed.

```
BenchmarkMemBufferSetGetLongKey
BenchmarkMemBufferSetGetLongKey/ART
BenchmarkMemBufferSetGetLongKey/ART-32           1000000               458.1 ns/op          2149 B/op          0 allocs/op
```

The common cases aren't that stable on my PC, and even the master code runs much slower than when the result in the description now (there's not much difference anyway).

The unsafe implementation runs faster, primarily due to better CPU pipeline utilization and more accurate prediction through chunk comparison. However, it seems to only work on little-endian machines, and I'm not sure if we need to worry about big-endian architectures...