@ekexium 2024-10-30
> The unsafe implementation runs faster, the only concern is it seems to only work on little-endian machines, and I'm not sure if we need to worry about big-endian architectures...

We can check internal/goarch, and fallback to other methods for big-endian architectures.