@cfzjywxk 2024-11-01
> /hold
> 
> The client-go's test seems ok, but the memory is violated when running with tidb.
> 
> ```
> fatal error: checkptr: pointer arithmetic result points to invalid allocation
> 
> goroutine 21334 gp=0xc002051a40 m=12 mp=0xc00088aa08 [running]:
> runtime.throw({0x97a1a64?, 0x216ae1d?})
> 	GOROOT/src/runtime/panic.go:1067 +0x48 fp=0xc004e40608 sp=0xc004e405d8 pc=0x21dc848
> runtime.checkptrArithmetic(0xc005f7bc0f?, {0xc004e406a8, 0x1, 0x219e3e9?})
> 	GOROOT/src/runtime/checkptr.go:69 +0x9c fp=0xc004e40638 sp=0xc004e40608 pc=0x216afbc
> github.com/tikv/client-go/v2/internal/unionstore/art.longestCommonPrefixByChunk({0xc00603a9d8, 0x8, 0x18?}, {0xc004f97138, 0x9, 0x21e7bd9?}, 0x0)
> 	external/com_github_tikv_client_go_v2/internal/unionstore/art/art_node.go:388 +0xfc fp=0xc004e406c0 sp=0xc004e40638 pc=0x3db30dc
> github.com/tikv/client-go/v2/internal/unionstore/art.longestCommonPrefix({0xc00603a9d8, 0x8, 0x8}, {0xc004f97138, 0x9, 0x14}, 0x0)
> 	external/com_github_tikv_client_go_v2/internal/unionstore/art/art_node.go:355 +0x69 fp=0xc004e40710 sp=0xc004e406c0 pc=0x3db2f69
> github.com/tikv/client-go/v2/internal/unionstore/art.(*nodeBase).match(...)
> 	external/com_github_tikv_client_go_v2/internal/unionstore/art/art_node.go:314
> ```

@you06 Is this a new bug introduced in https://github.com/tikv/client-go/pull/1477?