@you06 2024-11-07
> @you06 Is this a new bug introduced in #1477?

#1477 introduced a bug in the `baseIter.next`.

The error `arithmetic checkptr: pointer arithmetic result points to invalid allocation` is caused by `unsafe.Pointer(uintptr(p1) + 8)`, we should use `unsafe.Add` instead.

The CI in TiDB can pass after fixing both bugs.